api/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
google-api-python-client==2.18.0
psycopg2-binary==2.8.5
celery==5.1.2
redis==3.5.3

